services/api/pyproject.toml
[project]
name = "financial-api"
version = "0.1.0"
dependencies = [
  "fastapi>=0.112",
  "uvicorn[standard]>=0.30",
  "python-dotenv>=1.0",
  "financial-core @ file:///${PROJECT_ROOT}/packages/core"
]
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"